Document and restructure a bit
